Implemented Undo/Redo. Refactored logger to be easier to use. Refactored event handling loop to use actions.

+++ viewModels/BrushSettings.cpp
@@ -18,7 +18,7 @@ double BrushSettings::GetLineWidthAsDouble() const
 
 float BrushSettings::GetLineWidthAsFloat() const
 {
-	return lineWidth;
+	return (float)lineWidth;
 }
 
 void BrushSettings::SetLineWidth(const double lineWidth)
